разбивает исходный js код на отдельные модули
